Extract channel management to dedicated class
